pyproject.toml
[project]
name = "take-home"
version = "0.1.0"
description = "Barebones CLI chat agent built on LangChain Deep Agents"
requires-python = ">=3.13"
dependencies = [
    "deepagents",
    "langchain",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-google-genai",
    "python-dotenv",
]

[project.scripts]
chat = "agent.cli:main"
harness = "agent.harness:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
testpaths = ["evals"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
